+++ src/client/common/cy/events/contextMenu.js
